requirements.txt
streamlit
pillow
reportlab
streamlit-authenticator
gspread
oauth2client
